Split config resolution into another file
