+++ core/macrostrat/core/config.py
@@ -11,7 +11,7 @@ from sqlalchemy.engine import make_url
 from sqlalchemy.engine.url import URL
 from toml import load as load_toml
 
-from .utils import find_macrostrat_config
+from .utils import find_macrostrat_config, path_list_resolver, convert_to_string
 
 log = get_logger(__name__)
 
@@ -20,6 +20,7 @@ class MacrostratConfig(Dynaconf):
     """Macrostrat config manager that reads from a TOML file"""
 
     config_file: Path
+    srcroot: Path
 
     def __init__(self, *args, **kwargs):
         cfg = find_macrostrat_config()
@@ -39,6 +40,10 @@ class MacrostratConfig(Dynaconf):
         self.config_file = None
         if cfg is not None:
             self.config_file = Path(cfg)
+
+        # TODO: this enables sketchy behavior and tight coupling and should be removed.
+        # However it is a useful hack for now
+        self.srcroot = Path(__file__).parent.parent.parent.parent
 
     def all_environments(self):
         # Parse out top-level headers from TOML file
@@ -63,62 +68,17 @@ class MacrostratConfig(Dynaconf):
 settings = MacrostratConfig()
 
 
-def convert_to_string(value):
-    if value is None:
-        return None
-    return str(value)
-
-
-def list_of_paths(value):
-    if value is None:
-        return []
-    if isinstance(value, str):
-        return [Path(value)]
-    if isinstance(value, list):
-        return [Path(v) for v in value]
-    raise ValueError(f"Expected a string or list of strings, got {type(value)}")
-
-
-def list_of_resolved_paths(value, *, require_directory=False, require_file=False):
-    _paths = list_of_paths(value)
-    element_type = "file or directory"
-    if require_directory and require_file:
-        raise ValueError("Cannot require both files and directories")
-    if require_directory:
-        element_type = "directory"
-    if require_file:
-        element_type = "file"
-
-    _out = []
-    for p in _paths:
-        p1 = p.expanduser()
-        if not p1.is_absolute():
-            # Resolve relative to config file
-            if settings.config_file is not None:
-                p1 = settings.config_file.parent / p1
-        if not p1.exists():
-            raise ValueError(f"{p1} is not a valid {element_type}")
-        if require_directory and not p1.is_dir():
-            raise ValueError(f"{p1} is not a {element_type}")
-        if require_file and not p1.is_file():
-            raise ValueError(f"{p1} is not a {element_type}")
-        _out.append(p1.resolve())
-    return _out
-
-
-def list_of_directories(value):
-    return list_of_resolved_paths(value, require_directory=True)
-
-
-def list_of_files(value):
-    return list_of_resolved_paths(value, require_file=True)
-
-
 settings.validators.register(
     # `must_exist` is causing huge problems
     Validator("COMPOSE_ROOT", cast=Path),
-    Validator("env_files", cast=list_of_files, default=None),
-    Validator("script_dirs", cast=list_of_directories, default=None),
+    Validator(
+        "env_files", cast=path_list_resolver(settings, require_file=True), default=None
+    ),
+    Validator(
+        "script_dirs",
+        cast=path_list_resolver(settings, require_directory=True),
+        default=None,
+    ),
     Validator("pg_database", cast=convert_to_string, default=None),
     # Backend information. We could potentially infer this from other environment variables
     Validator("backend", default="kubernetes", cast=BackendType),
@@ -222,8 +182,6 @@ if mapbox_token := getattr(settings, "mapbox_token", None):
 if secret_key := getattr(settings, "secret_key", None):
     environ["SECRET_KEY"] = secret_key
 
-# Path to the root of the Macrostrat repository
-settings.srcroot = Path(__file__).parent.parent.parent.parent
 
 environ["MACROSTRAT_ROOT"] = str(settings.srcroot)
